Cambios Decimal

+++ Tintoreria/Core/Categoria.cs
@@ -7,7 +7,7 @@
         public int  IdCategoria { get; set; }
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
-        public double Precio { get; set; }
+        public decimal Precio { get; set; }
 
         #endregion
